Don't use the shell to invoke commands
Generally speaking, is very bad to execute commands using a shell, as
all sort of attacks can be done by injecting further commands using
shell tricks. It is also more prone to errors, as if an user names a
branch something like "x*" it could have unpredictable results when used
as part of a shell command.

Unless something is really needed from the shell, a process should be
spawn instead, and if a shell is really needed, the command passed
should be properly escaped to make sure the shell doesn't do crazy shit.

It is also more wasteful as it will spawn an extra process (the shell)
if it's unneeded.

This also removed the `Redirect.all` because it's the default and avoid
stripping the mail contents in tests, as is not needed anymore and
contents should match exactly.

Fixes #111.

test: Don't strip e-mail contents

E-mail contents shouldn't be stripped, they should match exactly.