Change release notes format to include summary line
In the old format, the "heading" for each change was the
names of the affected modules. This makes the compiled release
notes difficult to read, as the actual changes don't stand out.
The new format puts a summary of each change as the header,
making the whole file easier for humans to quickly parse.