AutoPR: Fix when versions have the same SHA
Fixes #172

When multiple tags point to the same SHA commit, we would detect the older
version for the SHA when detecting the current version.

This is now changed so that the latest version of multiple sha matches will be
chosen.

Additionally, a check was added if the SHA is equal to the version we want to
update to.