Add consistent version headers to release notes
Release notes for a particular tag are missing a heading for the main
release notes (the ones that are not inherited). To make notes more
consistent, we always add a header with the specific versions the following
notes are, so inherited notes will also get this very same header.

While at it, we stop passing around the inherited notes to
buildReleaseNotes(), as they are required also outside of that function.