Use more efficient way to check for PRs
We used to check for existing update PRs by getting all the recent PRs
of all the repositories and then name-matching them to our pattern.

This was error prone as a user could simply have renamed them and it is
inefficient as we fetch lots of PRs we don't care about.

This method was now changed to fetch the PRs based on the source branch
which always has the same name for us.